+++ packages/vue-cli-plugin-docgen/generator/template/src/components/AppButton.vue
@@ -54,7 +54,7 @@ export default {
 }
 </script>
 
-<docs>
+<docs lang="md">
 This button is amazing, use it responsibly.
 
 ## Examples
@@ -69,7 +69,7 @@ Ugly button with pink font and blue background:
 
 ```jsx
 <app-button color="pink" background="blue">
-  Ugly button
+	Ugly button
 </app-button>
 ```
 
@@ -77,7 +77,7 @@ Button containing custom tags:
 
 ```jsx
 <app-button>
-  Text with <b>bold</b>
+	Text with <b>bold</b>
 </app-button>
 ```
 </docs>