+++ packages/vue-docgen-api/src/parseSFC.ts
@@ -8,7 +8,7 @@ import parseTemplate, { Handler as TemplateHandler } from './parse-template'
 import Documentation from './Documentation'
 import { ParseOptions } from './parse'
 import parseScript from './parse-script'
-import defaultScriptHandlers, { preHandlers } from './script-handlers'
+import makePathResolver from './utils/makePathResolver'
 
 const read = promisify(readFile)
 
@@ -19,6 +19,9 @@ export default async function parseSFC(
 ): Promise<Documentation[]> {
 	let documentation = initialDoc
 
+	// create a custom path resolver to resolve webpack aliases
+	const pathResolver = makePathResolver(path.dirname(opt.filePath), opt.alias, opt.modules)
+
 	// use padding so that errors are displayed at the correct line
 	const { descriptor: parts } = cacher(() => parseComponent(source, { pad: 'line' }), source)
 
@@ -26,16 +29,20 @@ export default async function parseSFC(
 	if (parts.template) {
 		const extTemplSrc = parts?.template?.attrs?.src
 
-		const extTemplSource =
-			extTemplSrc && typeof extTemplSrc === 'string' && extTemplSrc.length
-				? await read(path.resolve(path.dirname(opt.filePath), extTemplSrc), {
+		if (extTemplSrc && typeof extTemplSrc === 'string') {
+			const extTemplSrcAbs = pathResolver(extTemplSrc)
+
+			const extTemplSource = extTemplSrcAbs
+				? await read(extTemplSrcAbs, {
 						encoding: 'utf-8'
 				  })
-				: // if we don't have a content to the binding, use empty string
-				  ''
+				: // if we don't have a content to bind
+				  // leave the template alone
+				  false
 
-		if (extTemplSource.length) {
-			parts.template.content = extTemplSource
+			if (extTemplSource) {
+				parts.template.content = extTemplSource
+			}
 		}
 		const addTemplateHandlers: TemplateHandler[] = opt.addTemplateHandlers || []
 
@@ -44,24 +51,31 @@ export default async function parseSFC(
 		parseTemplate(parts.template, documentation, [...templateHandlers, ...addTemplateHandlers], opt)
 	}
 
-	const extSrc = (parts && parts.script && parts.script.attrs
-		? parts.script.attrs.src
-		: '') as string
-	const extSource =
-		extSrc && extSrc.length
-			? await read(path.resolve(path.dirname(opt.filePath), extSrc), {
+	let scriptSource = parts.script ? parts.script.content : undefined
+
+	const extSrc = parts && parts.script && parts.script.attrs ? parts.script.attrs.src : false
+
+	if (extSrc && typeof extSrc === 'string') {
+		const extSrcAbs = pathResolver(extSrc)
+
+		const extSource = extSrcAbs
+			? await read(extSrcAbs, {
 					encoding: 'utf-8'
 			  })
 			: ''
+		if (extSource.length) {
+			scriptSource = extSource
+			opt.lang =
+				(parts.script && parts.script.attrs && /^tsx?$/.test(parts.script.attrs.lang as string)) ||
+				/\.tsx?$/i.test(extSrc)
+					? 'ts'
+					: 'js'
+		}
+	}
 
-	const scriptSource = extSource.length
-		? extSource
-		: parts.script
-		? parts.script.content
-		: undefined
 	opt.lang =
-		(parts.script && parts.script.attrs && parts.script.attrs.lang === 'ts') ||
-		/\.tsx?$/i.test(extSrc)
+		(parts.script && parts.script.attrs && /^tsx?$/.test(parts.script.attrs.lang as string)) ||
+		(typeof extSrc === 'string' && /\.tsx?$/i.test(extSrc))
 			? 'ts'
 			: 'js'
 
@@ -77,20 +91,8 @@ export default async function parseSFC(
 		}
 	}
 
-	const scriptHandlers = opt.scriptHandlers || [
-		...defaultScriptHandlers,
-		...(opt.addScriptHandlers || [])
-	]
-
 	const docs: Documentation[] = scriptSource
-		? (await parseScript(
-				scriptSource,
-				opt.scriptPreHandlers || preHandlers,
-				scriptHandlers,
-				opt,
-				documentation,
-				initialDoc !== undefined
-		  )) || []
+		? (await parseScript(scriptSource, opt, documentation, initialDoc !== undefined)) || []
 		: // if there is only a template return the template's doc
 		documentation
 		? [documentation]