+++ packages/vue-docgen-cli/src/templates/slots.ts
@@ -2,19 +2,38 @@ import { SlotDescriptor } from 'vue-docgen-api'
 import { mdclean } from './utils'
 import { SubTemplateOptions } from '../compileTemplates'
 
+const formatBindings: (slot: SlotDescriptor['bindings']) => string = bindings => {
+	if (!bindings) {
+		return ''
+	}
+	return bindings
+		.map(binding => {
+			const { name, description, type } = binding
+			if (!type) {
+				return ''
+			}
+			return `**${name}** \`${
+				type.name === 'union' && type.elements
+					? type.elements.map(({ name: insideName }) => insideName).join(', ')
+					: type.name
+			}\` - ${description}`
+		})
+		.join('\n')
+}
+
 export default (slots: SlotDescriptor[], opt: SubTemplateOptions = {}): string => {
 	return `
 ${opt.isSubComponent || opt.hasSubComponents ? '#' : ''}## Slots
 
-| Name          | Description  | Bindings |
-| ------------- | ------------ | -------- |
-${slots
-	.map(slot => {
-		const { description: d, bindings, name } = slot
-		const readableBindings =
-			bindings && Object.keys(bindings).length ? JSON.stringify(bindings, null, 2) : '' // serialize bindings to display them ina readable manner
-		return `| ${mdclean(name)} | ${mdclean(d || '')} | ${mdclean(readableBindings)} |` // remplace returns by <br> to allow them in a table cell
-	})
-	.join('\n')}
+  | Name          | Description  | Bindings |
+  | ------------- | ------------ | -------- |
+  ${slots
+		.map(slot => {
+			const { description, bindings, name } = slot
+			const readableBindings = bindings ? `${formatBindings(bindings)}` : ''
+
+			return `| ${mdclean(name)} | ${mdclean(description || '')} | ${mdclean(readableBindings)} |` // replace returns by <br> to allow them in a table cell
+		})
+		.join('\n')}
 `
 }