+++ packages/vue-styleguidist/src/bin/styleguidist.ts
@@ -4,6 +4,7 @@ import minimist from 'minimist'
 import kleur from 'kleur'
 import createLogger from 'glogg'
 import StyleguidistError from 'react-styleguidist/lib/scripts/utils/error'
+import { version as vueVersion } from 'vue/package.json'
 import { SanitizedStyleguidistConfig } from '../types/StyleGuide'
 import getConfig from '../scripts/config'
 import consts from '../scripts/consts'
@@ -35,6 +36,7 @@ process.on('uncaughtException', (err: any) => {
 	process.exit(1)
 })
 
+// eslint-disable-next-line no-import-assign
 const vueVersion = require('vue/package.json').version as string
 
 const VUE_VERSION = 3