+++ packages/vue-styleguidist/src/client/utils/processSections.js
@@ -15,7 +15,7 @@ export default function processSections(
 ) {
 	return sections.map(section => {
 		const options = {
-			useRouterLinks,
+			useRouterLinks: Boolean(useRouterLinks && section.name),
 			useHashId: section.sectionDepth === 0,
 			hashPath: [...hashPath, section.name ? section.name : '-']
 		}