+++ packages/vue-styleguidist/src/scripts/config.ts
@@ -19,7 +19,9 @@ const CONFIG_FILENAME = 'styleguide.config.js'
  */
 export default function getConfig(
 	configParam: string | StyleguidistConfig | { serverPort?: string | number },
-	update?: (conf: SanitizedStyleguidistConfig | {}) => SanitizedStyleguidistConfig
+	update?: (
+		conf: SanitizedStyleguidistConfig | { serverPort?: string | number }
+	) => SanitizedStyleguidistConfig
 ): Promise<SanitizedStyleguidistConfig> | SanitizedStyleguidistConfig {
 	let configFilepath
 	let config:
@@ -58,7 +60,9 @@ export default function getConfig(
 function postTreatConfig(
 	configDir: string,
 	config: StyleguidistConfig | { serverPort?: string | number },
-	update?: (conf: SanitizedStyleguidistConfig | {}) => SanitizedStyleguidistConfig
+	update?: (
+		conf: SanitizedStyleguidistConfig | { serverPort?: string | number }
+	) => SanitizedStyleguidistConfig
 ): SanitizedStyleguidistConfig {
 	if (update) {
 		config = update(config)